#8 [chore] Move customview to view/
