+++ app/src/main/res/layout/fragment_welcome.xml
@@ -44,7 +44,7 @@
                 app:layout_constraintStart_toStartOf="parent"
                 app:layout_constraintTop_toBottomOf="@id/image_view_welcome_image" />
 
-            <com.jjbaksa.jjbaksa.jj.JjEditText
+            <com.jjbaksa.jjbaksa.view.JjEditText
                 android:id="@+id/jj_edit_text_welcome_name"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
